Remove hacky testing of commonmark parsing testing
The commonmark parsing tests are already done by markdownit,
as suche we only need to test serializing. And of cause parsing of
markdown extensions we added.

The current "CommonMark" testing also had quite a lot of hacks
as even if we are commonmark compatible we do not yield the same html
representation (html is only a interim stage to tiptap).

Signed-off-by: Ferdinand Thiessen <rpm@fthiessen.de>